refactor: split common/common.go into tag, logicaltype, and utilities
Extract tag parsing into tag.go, logical type helpers into logicaltype.go,
and move matching tests to dedicated test files.
